Add files via upload
This code is used to set up a telegram bot, identify the accident-prone area, and display the data on the display attached to the device. Telegram bot is used to access the device remotely from any place and perform actions such as knowing the vehicle's current location, starting on/off the vehicle, etc. Accident-prone area detection is used to warn the driver in case he/she is entering accident prone area.